// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/control-calidad/ActaPendienteControlCalidad.ts
export interface ActaPendienteControlCalidad {
	idActa: number;
    mesa: string;
	copia: string;
	digitoChequeo: string;
	codigoEleccion: string;
	nombreEleccion: string;
	cvas: number;
	ubigeoDepa: string;
	ubigeoProv: string;
	ubigeoNombre: string;
	idArchivoActaEscrutinio: number;
	idArchivoActaInstalacionSufragio: number;
}