// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/control-calidad/DataPaso2.ts
export interface DataPaso2 {
    idImgFirmaPresidenteEsc: number;
    idImgFirmaSecretarioEsc: number;
    idImgFirmaTercerMiembroEsc: number;

    idImgFirmaPresidenteIns: number;
    idImgFirmaSecretarioIns: number;
    idImgFirmaTercerMiembroIns: number;
    
    idImgFirmaPresidenteSuf: number;
    idImgFirmaSecretarioSuf: number;
    idImgFirmaTercerMiembroSuf: number;
    
    idImgObsEscrutinio: number;
    idImgObsInstalacion: number;
    idImgObsSufragio: number;
    
    actaConFirma: boolean;
}

export interface SrcImagenesPaso2 {
    srcImgFirmaPresidenteEsc: string;
    srcImgFirmaSecretarioEsc: string;
    srcImgFirmaTercerMiembroEsc: string;

    srcImgFirmaPresidenteIns: string;
    srcImgFirmaSecretarioIns: string;
    srcImgFirmaTercerMiembroIns: string;
    
    srcImgFirmaPresidenteSuf: string;
    srcImgFirmaSecretarioSuf: string;
    srcImgFirmaTercerMiembroSuf: string;
    
    srcImgObsEscrutinio: string;
    srcImgObsInstalacion: string;
    srcImgObsSufragio: string;
    
    actaConFirma: boolean;
}